Another example of using vitae
The codes in this repository recreate headers and styles from Awesome CV while allowing more direct control on the .tex code that runs underneath. This is achieved by directing the output of your rmarkdown file to `awesome.tex` which controls format and styles. 

```
output:
  pdf_document:
    latex_engine: xelatex
    template: "awesome.tex"
```
The awesome.tex is a simplified version of awesome-cv.cls.